+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## dev-master
 
+* Table: accumulate INSERT
 * #34: Fixed `?w` for PgSQL (argument of WHERE must be type boolean)
 * #31, #32: Fix error messages for PgSQL
-* Table: accumulate INSERT
+* #37: Fixed exception for named data
 
 ## 2.2.1 (03.07.2014)